Rewrite args
